fix(enc): type the frame header's dictionary_id as u32
The Dictionary_ID wire field has no class wider than 4 bytes, but the
pub FrameHeader carried Option<u64> — a > u32::MAX id constructed by any
external caller reached the raw panic!() arm in descriptor(). Narrow the
field to u32 (header_id and compress_with_state_dictionary follow; the
decoder side already parses u32), so the unserializable ids are rejected
at compile time; the dead 0=>0 arm goes and the impossible tail becomes
a documented unreachable. Boundary test pins all three wire classes.

+++ crates/zstdx/src/encoding/dictionary.rs
@@ -88,9 +88,10 @@ impl EncDictionary {
 
     /// The dictID to declare in the frame header: a zero id means "no id"
     /// (the format's null dictID) — entropy seeding is skipped there too,
-    /// so the frame stays decodable without the dictionary.
-    pub fn header_id(&self) -> Option<u64> {
-        (self.id != 0).then_some(self.id as u64)
+    /// so the frame stays decodable without the dictionary. u32 caps the
+    /// value at the Dictionary_ID field's 4-byte wire limit.
+    pub fn header_id(&self) -> Option<u32> {
+        (self.id != 0).then_some(self.id)
     }
 }
